feat(recipes): migrate recipes to TypeScript and update dependencies
- Added TypeScript support for the following recipes: simon_says, speech_to_text, text_to_speech, and tj_the_translator.
- Converted JavaScript files (.mjs) to TypeScript files (.mts).
- Updated package.json to use tsx for running scripts.
- Created mise.toml and recipe.toml files for each recipe with necessary configurations.
- Removed old ESLint configuration files and replaced them with new configurations using @eslint/js and @typescript-eslint.
- Updated TypeScript configuration (tsconfig.json) for each recipe to target ES2020 and use NodeNext module resolution.
- Removed deprecated or unnecessary files and configurations.
