+++ recipes/chat_with_tj/index.ts
@@ -1,0 +1,139 @@
+/**
+ * Copyright 2024-2025 IBM Corp. All Rights Reserved.
+ * Copyright 2026-present TJBot Contributors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import TJBot from 'tjbot';
+import { WatsonXAI } from '@ibm-cloud/watsonx-ai';
+
+const BASE_PROMPT = `
+You are TJBot, a friendly and helpful social robot made out of cardboard.
+You are having a conversation with a human.
+You provide friendly and helpful responses to everything the human says.
+You respond to their questions in a professional manner.
+You never use inappropriate language like swear words or hate speech.
+You aim to be delightful and energetic in your responses.
+If you don't know the answer to a question, you respond truthfully that you do not know.
+`;
+
+// read recipe-specific config
+const config = TJBot.getRecipeConfig();
+
+// these are the hardware capabilities that TJ needs for this recipe
+const hardware = [
+    TJBot.Hardware.MICROPHONE,
+    TJBot.Hardware.SPEAKER,
+];
+
+let hasLED = false;
+if (config.useNeoPixelLED) {
+    hardware.push(TJBot.Hardware.LED_NEOPIXEL);
+    hasLED = true;
+}
+if (config.useCommonAnodeLED) {
+    hardware.push(TJBot.Hardware.LED_COMMON_ANODE);
+    hasLED = true;
+}
+
+// create an instance of the watsonx.ai service
+const wxai = WatsonXAI.newInstance({
+    serviceUrl: config.serviceUrl,
+    version: config.serviceVersion,
+});
+
+// keep track of the conversational history
+let conversationHistory = '';
+
+// instantiate our TJBot!
+const tj = await TJBot.getInstance().initialize({
+    hardware: hardware
+});
+
+// ready!
+console.log('TJBot is ready for conversation!');
+console.log("Say 'stop' or press ctrl-c to exit this recipe.");
+
+while (true) {
+    console.log('👂 listening...');
+
+    if (hasLED) {
+        tj.shine('green');
+    }
+
+    let msg = await tj.listen();
+
+    if (hasLED) {
+        tj.pulse('orange');
+    }
+
+    if (msg === undefined || msg === '') {
+        continue;
+    }
+
+    if (msg.toLowerCase() === 'stop') {
+        console.log('Goodbye!');
+        process.exit(0);
+    }
+
+    // strip out %HESITATION
+    msg = msg.replaceAll('%HESITATION', '');
+
+    // build the prompt
+    const prompt = `
+${BASE_PROMPT}
+
+Conversation summary:
+${conversationHistory}
+
+Conversation:
+
+H: ${msg}
+AI: `;
+
+    const params = {
+        input: prompt,
+        modelId: config.modelId,
+        projectId: config.projectId,
+        parameters: {
+            decoding_method: config.modelDecodingMethod || 'greedy',
+            temperature: config.modelTemperature || 0.7,
+            random_seed: config.modelRandomSeed || 42,
+            min_new_tokens: config.modelMinNewTokens || 0,
+            max_new_tokens: config.modelMaxNewTokens || 200,
+            stop_sequences: ['.'],
+            repetition_penalty: config.modelRepetitionPenalty || 1.0,
+        },
+    };
+
+    try {
+        const response = await wxai.generateText(params);
+        const text = response.result.results[0].generated_text;
+
+        console.log(`👩‍💻 > ${msg}`);
+        console.log(`🤖 > ${text}`);
+
+        console.log('🗯️ speaking...');
+        if (hasLED) {
+            tj.pulse('yellow');
+        }
+        await tj.speak(text);
+        console.log('🗯️ speaking finished');
+
+        // add to the conversation history
+        conversationHistory += `Human: ${msg}\n AI: ${text}\n\n`;
+    } catch (err) {
+        console.warn(`⚠️ > ${err}`);
+    }
+}
