+++ recipes/hello_tjbot/package.json
@@ -5,9 +5,9 @@
   "license": "Apache-2.0",
   "author": "TJBot Community Edition <https://tjbot-ce.github.io>",
   "type": "module",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
-    "start": "node index.js",
+    "start": "tsx index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
@@ -15,7 +15,11 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.39.2",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
     "eslint": "^9.39.2",
-    "globals": "^17.0.0"
+    "globals": "^17.0.0",
+    "tsx": "^4.19.0",
+    "typescript": "^5.6.0"
   }
 }
