+++ recipes/hello_tjbot/recipe.toml
@@ -1,0 +1,10 @@
+[recipe]
+name = "hello_tjbot"
+description = "Hello world, TJBot style."
+version = "3.0.0"
+author = "TJBot Community Edition <https://tjbot-ce.github.io>"
+license = "Apache-2.0"
+
+# Recipe-specific configuration
+# Add any custom configuration options your recipe needs here.
+[config]
