--- recipes/simon_says/.eslintrc.json
@@ -1,20 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "es2021": true
-    },
-    "extends": [
-        "airbnb-base"
-    ],
-    "parserOptions": {
-        "ecmaVersion": "latest",
-        "sourceType": "module"
-    },
-    "rules": {
-        "indent": ["error", 4],
-        "max-len": "off",
-        "no-console": "off",
-        "no-underscore-dangle": "off",
-        "import/extensions": "off"
-    }
-}
