+++ recipes/simon_says/recipe.toml
@@ -1,0 +1,10 @@
+[recipe]
+name = "simon_says"
+description = "TJBot plays Simon Says, powered by IBM watsonx Assistant"
+version = "3.0.0"
+author = "TJBot <https://ibmtjbot.github.io>"
+license = "Apache-2.0"
+
+# Recipe-specific configuration
+# Add any custom configuration options your recipe needs here.
+[config]
