+++ recipes/speech_to_text/recipe.toml
@@ -1,0 +1,10 @@
+[recipe]
+name = "speech_to_text"
+description = "Change the color of TJBot's LED with your voice!"
+version = "3.0.0"
+author = "TJBot <https://ibmtjbot.github.io>"
+license = "Apache-2.0"
+
+# Recipe-specific configuration
+# Add any custom configuration options your recipe needs here.
+[config]
