+++ recipes/text_to_speech/recipe.toml
@@ -1,0 +1,10 @@
+[recipe]
+name = "text_to_speech"
+description = "Make your TJBot speak!"
+version = "3.0.0"
+author = "TJBot <https://ibmtjbot.github.io>"
+license = "Apache-2.0"
+
+# Recipe-specific configuration
+# Add any custom configuration options your recipe needs here.
+[config]
