+++ recipes/tj_the_translator/package.json
@@ -16,9 +16,9 @@
   "license": "Apache-2.0",
   "author": "TJBot <https://ibmtjbot.github.io>",
   "type": "module",
-  "main": "index.mjs",
+  "main": "index.mts",
   "scripts": {
-    "start": "sudo node index.mjs",
+    "start": "sudo tsx index.mts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
@@ -26,8 +26,12 @@
     "tjbot": "file:../../../node-tjbotlib"
   },
   "devDependencies": {
-    "eslint": "^8.57.1",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.31.0"
+    "@eslint/js": "^9.39.2",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
+    "eslint": "^9.39.2",
+    "globals": "^17.0.0",
+    "tsx": "^4.19.0",
+    "typescript": "^5.6.0"
   }
 }
